Split sides_t into its own module
